reorganized pages and component to a general standard
