+++ ui/src/App.js
@@ -2,9 +2,10 @@ import 'bootstrap/dist/css/bootstrap.min.css'
 import React from 'react'
 import Login from "./pages/Login.js"
 import Home from "./pages/Home.js"
-import Post from "./pages/Post.js"
+import Post from "./components/Landing/Post.js"
 import Request from "./pages/Request.js"
-import NavBar from "./components/Navbar"
+import RequestCreate from "./pages/RequestCreate.js"
+import NavBar from "./components/Interface/Navbar"
 import {Routes, Route} from "react-router-dom"
 
 
@@ -17,6 +18,7 @@ function App() {
         <Route path="/home" element={<Home />} />
         <Route path="/home/post/:postID" element={<Post />} />
         <Route path="/request" element={<Request />} />
+        <Route path="/create" element={<RequestCreate />} />
       </Routes>
     </div>
   );
