+++ ui/src/components/Interface/Navbar.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import {useLocation} from 'react-router-dom'
 import {Navbar,  Container, Button} from 'react-bootstrap'
-import "./components.css"
-import logoPath from "../images/overbrook.png"
+import "./style.css"
+import logoPath from "../../images/overbrook.png"
 
 export default function NavBar() {
