+++ ui/src/components/Interface/style.css
@@ -1,6 +1,6 @@
 .priority-announcement {
     margin-top: 10px;
-    padding: 50px;
+    padding: 20px;
 }
 
 .priority-announcement-text {
