+++ ui/src/components/RequestCreate/RequestCreateComp.js
@@ -1,0 +1,17 @@
+import Card from "react-bootstrap/Card";
+import Container from "react-bootstrap/Container";
+import Button from "react-bootstrap/Button";
+import "bootstrap/dist/css/bootstrap.min.css";
+
+import "../Interface/style.css"
+
+export default function RequestCreate() {
+  return (
+    <Container className="priority-announcement">
+      <Card className="shadow-sm">
+        <Card.Body>
+        </Card.Body>
+      </Card>
+    </Container>
+  );
+}
