+++ ui/src/pages/Home.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import PriorityAnnouncement from '../components/PriorityAnnouncement';
-import Announcements from '../components/Announcements';
+import PriorityAnnouncement from '../components/Landing/PriorityAnnouncement';
+import Announcements from '../components/Landing/Announcements';
 
 export default function Home() {
   return (
