+++ ui/src/pages/Request.js
@@ -1,14 +1,11 @@
 import "bootstrap/dist/css/bootstrap.min.css";
 import { Fragment, React } from "react";
-import Requests from "../components/Requests";
+import RequestComp from "../components/Request/RequestComp";
 
 export default function Request() {
   return (
     <Fragment>
-      <header>something</header>
-      <main>
-          <Requests />
-      </main>
+          <RequestComp />
     </Fragment>
   );
 }
