+++ ui/src/pages/RequestCreate.js
@@ -1,0 +1,12 @@
+import "bootstrap/dist/css/bootstrap.min.css";
+import { Fragment, React } from "react";
+import RequestCreateComp from "../components/RequestCreate/RequestCreateComp"
+
+
+export default function RequestCreate() {
+  return (
+    <Fragment>
+          <RequestCreateComp />
+    </Fragment>
+  );
+}
